fastapi/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pydantic[email]
boto3
bcrypt
PyJWT
fastapi-mail
requests
python-dotenv
